[feat] 카테고리 파베에 넘기기
