+++ src/components/CategoriesButton.jsx
@@ -30,24 +30,4 @@ const CategoriesButton = ({ filter, selected, handleFilterClick }) => {
   );
 };
 
-const CategoryButton = () => {
-  const filters = ["생활습관", "식습관", "취미", "환경"];
-
-  const handleFilterClick = (filter) => {
-  };
-
-  return (
-    <div className="flex gap-1">
-      {filters.map((filter) => (
-        <CategoriesButton
-          key={filter}
-          filter={filter}
-          selected={false}
-          handleFilterClick={handleFilterClick}
-        />
-      ))}
-    </div>
-  );
-};
-
-export default CategoryButton;
+export default CategoriesButton
